Change player class before equipping weapons
Before we equip weapons, we change the player's class to the default class of the weapon we are equipping.
This fixes classes having different reload times with the same weapons, as well as a crash with sniper when using a banner.
Note: Using sm_rndclass can still cause broken reload times